+++ src/components/ComponentTreeWithContext.tsx
@@ -1,0 +1,11 @@
+import {MessageWithContext} from "./MessageWithContext.tsx";
+import {IncrementCountWithContext} from "./IncrementCountWIthContext.tsx";
+
+export function ComponentTreeWithContext() {
+    return (
+        <div className="component-tree">
+            <IncrementCountWithContext />
+            <MessageWithContext />
+        </div>
+    );
+}
