+++ src/components/CounterWIthContext.tsx
@@ -1,0 +1,12 @@
+import {useContext} from "react";
+import {MyContext} from "../context.tsx";
+
+export function CounterWIthContext() {
+    const { clickCount } = useContext(MyContext);
+
+    return (
+        <div>
+            <p>Click count: {clickCount}</p>
+        </div>
+    );
+}
