+++ src/components/MessageWithContext.tsx
@@ -1,0 +1,10 @@
+import {CounterWIthContext} from "./CounterWIthContext.tsx";
+
+export function MessageWithContext() {
+    return (
+        <div>
+            <CounterWIthContext />
+            <p>Message</p>
+        </div>
+    );
+}
