+++ src/context.tsx
@@ -1,0 +1,6 @@
+import {createContext} from "react";
+
+export const MyContext = createContext({
+    clickCount: 0,
+    increment: () => {}
+});
